+++ src/App.css
@@ -27,7 +27,6 @@
 .App-link {
   color: #61dafb;
 }
-
 @keyframes App-logo-spin {
   from {
     transform: rotate(0deg);
@@ -35,4 +34,23 @@
   to {
     transform: rotate(360deg);
   }
+}
+.bg-gradient {
+  background: linear-gradient(rgba(245, 245, 245), rgba(245, 245, 245, 0));
+}.bgstyle {
+  /* background: linear-gradient(136deg, #1ad7c0, #8E44AD, #ff9b11, #1F3A93, #ddd8c6, #000000, #000000);  */
+  background: linear-gradient(
+    136deg,
+    #4484ce,
+    #1ad7c0,
+    #ff9b11,
+    #9b59b6,
+    #ff7f7f);
+
+  background-size: 1200% 1200%;
+  -webkit-animation: AnimationName 70s ease infinite;
+  -moz-animation: AnimationName 70s ease infinite;
+  -o-animation: AnimationName 70s ease infinite;
+  animation: AnimationName 70s ease infinite;
+  height: 100%;
 }
